style: move wrapper functions to separate file
